Fix second pow ambiguity in Script.cpp for PS3
Explicitly cast pow arguments to double in the second conversion loop in Script.cpp.

+++ RSDKv4/Script.cpp
@@ -2569,7 +2569,7 @@ void AppendIntegerToStringW(ushort *text, int value)
 
     v = 0;
     for (int i = cnt - 1; i >= 0; --i) {
-        v = value / pow(10, i);
+        v = value / (int)pow(10.0, i);
         v %= 10;
 
         int strValue = v + '0';
